+++ AppMetrics.sln
@@ -34,8 +34,6 @@ Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "App.Metrics.Concurrency.Fac
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "benchmarks", "benchmarks", "{A335612E-C4C6-4AF4-98A9-EB9FF40CA258}"
 EndProject
-Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "App.Metrics.Concurrency.Benchmarks", "benchmarks\App.Metrics.Concurrency.Benchmarks\App.Metrics.Concurrency.Benchmarks.xproj", "{E81258B1-385B-4399-9BD6-5BB123B966A7}"
-EndProject
 Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "App.Metrics.Extensions.Mvc", "src\App.Metrics.Extensions.Mvc\App.Metrics.Extensions.Mvc.xproj", "{1C13412E-E6A8-4617-9EC3-AEF40CEFA2F0}"
 EndProject
 Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "App.Metrics.Net452.Facts", "test\App.Metrics.Net452.Facts\App.Metrics.Net452.Facts.xproj", "{9921B544-64C6-477D-93F1-2916289022B6}"
@@ -82,10 +80,6 @@ Global
 		{66CF57E1-7D50-427A-8047-54C45A7FE90B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{66CF57E1-7D50-427A-8047-54C45A7FE90B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{66CF57E1-7D50-427A-8047-54C45A7FE90B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E81258B1-385B-4399-9BD6-5BB123B966A7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E81258B1-385B-4399-9BD6-5BB123B966A7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E81258B1-385B-4399-9BD6-5BB123B966A7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E81258B1-385B-4399-9BD6-5BB123B966A7}.Release|Any CPU.Build.0 = Release|Any CPU
 		{1C13412E-E6A8-4617-9EC3-AEF40CEFA2F0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{1C13412E-E6A8-4617-9EC3-AEF40CEFA2F0}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1C13412E-E6A8-4617-9EC3-AEF40CEFA2F0}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -122,7 +116,6 @@ Global
 		{EBFE6BC4-D1DE-40D4-8CBD-11EAC36DBA04} = {DD2E4647-7125-4CE3-9BA5-B45A26113A31}
 		{F9AA4862-3755-425F-8F33-B25772F43FFC} = {2D805782-756E-4C98-B22E-F502BEE95318}
 		{66CF57E1-7D50-427A-8047-54C45A7FE90B} = {DD2E4647-7125-4CE3-9BA5-B45A26113A31}
-		{E81258B1-385B-4399-9BD6-5BB123B966A7} = {A335612E-C4C6-4AF4-98A9-EB9FF40CA258}
 		{1C13412E-E6A8-4617-9EC3-AEF40CEFA2F0} = {2D805782-756E-4C98-B22E-F502BEE95318}
 		{9921B544-64C6-477D-93F1-2916289022B6} = {DD2E4647-7125-4CE3-9BA5-B45A26113A31}
 		{66ADEF00-7776-4208-932D-D3D2F7EADF52} = {2D805782-756E-4C98-B22E-F502BEE95318}
