--- benchmark-results/AtomicLongBenchMark-report-github.md
@@ -1,22 +1,0 @@
-``` ini
-
-BenchmarkDotNet=v0.10.0
-OS=Windows
-Processor=?, ProcessorCount=8
-Frequency=3312791 Hz, Resolution=301.8603 ns, Timer=TSC
-Host Runtime=.NET Core 4.6.24410.01, Arch=64-bit  [RyuJIT]
-GC=Concurrent Workstation
-dotnet cli version=1.0.0-preview2-1-003177
-Job Runtime(s):
-	.NET Core 4.6.24410.01, Arch=64-bit  [RyuJIT]
-
-Job=QuickRun  Jit=RyuJit  Platform=X64  
-Runtime=Core  LaunchCount=1  RunStrategy=ColdStart  
-TargetCount=5  UnrollFactor=1  WarmupCount=5  
-
-```
-    Method |       Mean |     StdErr |     StdDev |    Median |
----------- |----------- |----------- |----------- |---------- |
- Decrement | 36.3440 us | 35.8173 us | 80.0899 us | 0.3019 us |
-       Get | 36.2232 us | 35.3955 us | 79.1467 us | 0.3019 us |
- Increment | 36.8873 us | 36.2857 us | 81.1374 us | 0.3019 us |
