--- benchmark-results/PaddedAtomicLongBenchMark-report-github.md
@@ -1,22 +1,0 @@
-``` ini
-
-BenchmarkDotNet=v0.10.0
-OS=Windows
-Processor=?, ProcessorCount=8
-Frequency=3312791 Hz, Resolution=301.8603 ns, Timer=TSC
-Host Runtime=.NET Core 4.6.24410.01, Arch=64-bit  [RyuJIT]
-GC=Concurrent Workstation
-dotnet cli version=1.0.0-preview2-1-003177
-Job Runtime(s):
-	.NET Core 4.6.24410.01, Arch=64-bit  [RyuJIT]
-
-Job=QuickRun  Jit=RyuJit  Platform=X64  
-Runtime=Core  LaunchCount=1  RunStrategy=ColdStart  
-TargetCount=5  UnrollFactor=1  WarmupCount=5  
-
-```
-    Method |       Mean |     StdErr |     StdDev |    Median |
----------- |----------- |----------- |----------- |---------- |
- Decrement | 42.5019 us | 41.6738 us | 93.1855 us | 0.3019 us |
-       Get | 38.3363 us | 37.8851 us | 84.7136 us | 0.0000 us |
- Increment | 37.4910 us | 36.7389 us | 82.1508 us | 0.3019 us |
