--- benchmark-results/ThreadLocalLongAdderBenchMark-report-github.md
@@ -1,22 +1,0 @@
-``` ini
-
-BenchmarkDotNet=v0.10.0
-OS=Windows
-Processor=?, ProcessorCount=8
-Frequency=3312791 Hz, Resolution=301.8603 ns, Timer=TSC
-Host Runtime=.NET Core 4.6.24410.01, Arch=64-bit  [RyuJIT]
-GC=Concurrent Workstation
-dotnet cli version=1.0.0-preview2-1-003177
-Job Runtime(s):
-	.NET Core 4.6.24410.01, Arch=64-bit  [RyuJIT]
-
-Job=QuickRun  Jit=RyuJit  Platform=X64  
-Runtime=Core  LaunchCount=1  RunStrategy=ColdStart  
-TargetCount=5  UnrollFactor=1  WarmupCount=5  
-
-```
-    Method |       Mean |     StdErr |     StdDev |    Median |
----------- |----------- |----------- |----------- |---------- |
- Decrement | 43.9509 us | 43.1227 us | 96.4252 us | 0.3019 us |
-       Get | 41.4756 us | 39.3639 us | 88.0203 us | 1.8112 us |
- Increment | 38.8796 us | 38.0517 us | 85.0862 us | 0.3019 us |
