--- benchmarks/App.Metrics.Concurrency.Benchmarks/project.json
@@ -1,27 +1,0 @@
-{
-  "buildOptions": {
-    "allowUnsafe": true
-  },
-  "dependencies": {
-    "App.Metrics.Concurrency": { "target": "project" },
-    "xunit": "2.1.0",
-    "dotnet-test-xunit": "1.0.0-rc2-build10025",
-    "BenchmarkDotNet": "0.10.0",
-    "BenchmarkDotNet.Core": "0.10.0"
-  },
-  "frameworks": {
-    "netcoreapp1.1": {
-      "dependencies": {
-        "Microsoft.NETCore.App": {
-          "type": "platform",
-          "version": "1.1.0"
-        }
-      },
-      "imports": [
-        "dnxcore50",
-        "portable-net45+win8"
-      ]
-    }
-  },
-  "testRunner": "xunit"
-}
